Buttons in example grey when widgets removed
